Inventory system
Picking up and object and it appears in the inventory. It can be selected in-between the objects using the Dpad up and down buttons. When pressed O button, it should be destroyed from the inventory but it still does not appear in the scene. that needs to be checked out but the rest are fine.